pyproject.toml
[project]
name = "permit-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiosqlite>=0.21.0",
    "mcp>=1.2.1",
    "permit>=2.7.2",
    "httpx>=0.28.1",
    "python-dotenv>=1.0.1",
]

[tool.uv.workspace]
members = ["examples/food-ordering-system"]
